etl/requirements.txt
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.0.3
python-dotenv==1.0.0
apscheduler==3.10.4

